feat: port split-layout Lumo styles to CSS files
