+++ dev/split-layout.html
@@ -9,6 +9,7 @@
 
     <script type="module">
       import '@vaadin/split-layout';
+      import '@vaadin/vaadin-lumo-styles/lumo.css';
 
       document.querySelectorAll('.first').forEach(el => {el.textContent = new Array(1000).fill('a').join(' ');});
       document.querySelectorAll('.second').forEach(el => {el.textContent = new Array(1000).fill('b').join(' ');});
