+++ packages/split-layout/test/visual/lumo/split-layout.test.js
@@ -1,6 +1,8 @@
 import { fixtureSync } from '@vaadin/testing-helpers/dist/fixture.js';
 import { visualDiff } from '@web/test-runner-visual-regression';
-import '../../../theme/lumo/vaadin-split-layout.js';
+import '@vaadin/vaadin-lumo-styles/props.css';
+import '@vaadin/vaadin-lumo-styles/components/split-layout.css';
+import '../../../vaadin-split-layout.js';
 
 describe('split-layout', () => {
   let element;
