+++ packages/vaadin-lumo-styles/components.css
@@ -22,6 +22,7 @@
 @import './components/number-field.css';
 @import './components/password-field.css';
 @import './components/select.css';
+@import './components/split-layout.css';
 @import './components/text-area.css';
 @import './components/text-field.css';
 @import './components/time-picker.css';
