+++ packages/vaadin-lumo-styles/components/split-layout.css
@@ -1,0 +1,11 @@
+/**
+ * @license
+ * Copyright (c) 2017 - 2025 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+@import '../src/mixins/base-layer-reset.css' vaadin-split-layout;
+@import '../src/components/split-layout.css' vaadin-split-layout;
+
+html {
+  --vaadin-split-layout-css-inject: 1;
+}
