+++ packages/vaadin-lumo-styles/src/components/split-layout.css
@@ -1,0 +1,131 @@
+/**
+ * @license
+ * Copyright (c) 2017 - 2025 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+:host {
+  display: flex;
+  overflow: hidden !important;
+  transform: translateZ(0);
+}
+
+:host([hidden]) {
+  display: none !important;
+}
+
+:host([orientation='vertical']) {
+  flex-direction: column;
+}
+
+:host ::slotted(*) {
+  flex: 1 1 auto;
+  overflow: auto;
+  -webkit-overflow-scrolling: touch;
+}
+
+[part='splitter'] {
+  flex: none;
+  position: relative;
+  z-index: 1;
+  overflow: visible;
+  min-width: var(--lumo-space-s);
+  min-height: var(--lumo-space-s);
+  background-color: var(--lumo-contrast-5pct);
+  transition: 0.1s background-color;
+}
+
+:host(:not([orientation='vertical'])) > [part='splitter'] {
+  cursor: ew-resize;
+}
+
+:host([orientation='vertical']) > [part='splitter'] {
+  cursor: ns-resize;
+}
+
+[part='handle'] {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate3d(-50%, -50%, 0);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: var(--lumo-size-m);
+  height: var(--lumo-size-m);
+}
+
+[part='handle']::after {
+  content: '';
+  display: block;
+  --_handle-size: 4px;
+  width: var(--_handle-size);
+  height: 100%;
+  max-width: 100%;
+  max-height: 100%;
+  border-radius: var(--lumo-border-radius-s);
+  background-color: var(--lumo-contrast-30pct);
+  transition:
+    0.1s opacity,
+    0.1s background-color;
+}
+
+:host([orientation='vertical']) [part='handle']::after {
+  width: 100%;
+  height: var(--_handle-size);
+}
+
+/* Active style */
+[part='splitter']:active [part='handle']::after {
+  background-color: var(--lumo-contrast-50pct);
+}
+
+/* Small/minimal */
+:host([theme~='small']) > [part='splitter'] {
+  border-left: 1px solid var(--lumo-contrast-10pct);
+  border-top: 1px solid var(--lumo-contrast-10pct);
+}
+
+:host(:is([theme~='small'], [theme~='minimal'])) > [part='splitter'] {
+  min-width: 0;
+  min-height: 0;
+  background-color: transparent;
+}
+
+:host(:is([theme~='small'], [theme~='minimal'])) > [part='splitter']::after {
+  content: '';
+  position: absolute;
+  inset: -4px;
+}
+
+:host(:is([theme~='small'], [theme~='minimal'])) > [part='splitter'] > [part='handle'] {
+  left: calc(50% - 0.5px);
+  top: calc(50% - 0.5px);
+}
+
+:host(:is([theme~='small'], [theme~='minimal'])) > [part='splitter'] > [part='handle']::after {
+  opacity: 0;
+  --_handle-size: 5px;
+}
+
+:host(:is([theme~='small'], [theme~='minimal'])) > [part='splitter']:hover > [part='handle']::after,
+:host(:is([theme~='small'], [theme~='minimal'])) > [part='splitter']:active > [part='handle']::after {
+  opacity: 1;
+}
+
+/* Hover style */
+@media (any-hover: hover) {
+  [part='splitter']:hover [part='handle']::after {
+    background-color: var(--lumo-contrast-40pct);
+  }
+}
+
+@media (forced-colors: active) {
+  [part~='splitter'] {
+    outline: 1px solid;
+  }
+
+  [part~='handle']::after {
+    background-color: AccentColor !important;
+    forced-color-adjust: none;
+  }
+}
